+++ resources/views/contact.blade.php
@@ -2,11 +2,10 @@
 
     <div class="container py-3">
         <div class="col-md-50 max">
-            <h1>{{ $title }}</h1>
-
-            <x-contactit-contact>
-                {{-- remove top margin if adding $slot content --}}
-            </x-contactit-contact>
+            <h1>Contact Us</h1>
+            <div class="bx">
+                <livewire:contact/>
+            </div>
         </div>
     </div>